style: ran prettier on src folder

+++ src/handlers/events/guildMemberRemove.ts
@@ -3,35 +3,30 @@ import { Interpreter } from "../../core"
 import { InviteSystem } from "../../structures/InviteSystem"
 import { DiscordEventHandler } from "../../structures/DiscordEventHandler"
 
-export default new DiscordEventHandler(
-    {
-        name: "guildMemberRemove",
-        version: "1.0.1",
-        description: "This event is fired when a member leaves, is kicked or banned from a guild",
-        listener: async function(m) {
-            if (this.options.useInviteSystem)
-                InviteSystem.deleteInviter(m)
+export default new DiscordEventHandler({
+    name: "guildMemberRemove",
+    version: "1.0.1",
+    description: "This event is fired when a member leaves, is kicked or banned from a guild",
+    listener: async function (m) {
+        if (this.options.useInviteSystem) InviteSystem.deleteInviter(m)
 
-            const commands = this.commands.get("guildMemberRemove")
-    
-            for (const command of commands) {
-                Interpreter.run({
-                    obj: m as GuildMember,
-                    command,
-                    client: this,
-                    states: {
-                        member: {
-                            old: m as GuildMember,
-                            new: m as GuildMember
-                        }
+        const commands = this.commands.get("guildMemberRemove")
+
+        for (const command of commands) {
+            Interpreter.run({
+                obj: m as GuildMember,
+                command,
+                client: this,
+                states: {
+                    member: {
+                        old: m as GuildMember,
+                        new: m as GuildMember,
                     },
-                    data: command.compiled.code,
-                    args: []
-                })
-            }
-        },
-        intents: [
-            "GuildMembers"
-        ]
-    }
-)
+                },
+                data: command.compiled.code,
+                args: [],
+            })
+        }
+    },
+    intents: ["GuildMembers"],
+})
